Add CI
Uses stackage-curator tool to unpack and build every package in the
snapshot. Attempts to leverage caching to avoid building unchanged
snapshots.

This may just be too slow to be useful. An alternative would be to
handle-specify a package that uses some set of dependencies in the
snapshot and build that. That would trade coverage for time.

We'll see.