Add CI
Creates a small test project that requires all the dependencies across
our main apps as of now. This will require a bit of maintenance (which
will likely just not happen), but does add enough isolated coverage that
it may be worth it.

For a failed attempt at getting actual coverage, see #10.