Simplify testing strategy
Before this, we would rely on a cache hit to ensure we don't test every
unchanged snapshot on every single PR. GitHub expires caches, which is
reasonable, and on a slow-moving project such as this one, that may
happen frequently causing a lot of building.

It seems like a good 80/20 to just always test the most recent snapshot.
Since snapshots on main must be considered immutable, any PR is all but
guaranteed to be changing the most recent snapshot only.