Create lts-18.25 Snapshot (#30)
* Create lts-18.25 Snapshot

* Update the versions without correct hash

* Update `freckle-app` hash

* Update `dhall` hash

* Rev aws-xray-client-persistent

This should fix the test issue. Will update hash in next commit

* Bump dhall

Didn't I just do this?

* Bump `aws-xray-client-persistent`

* Fix `aws-xray-client-persistent` hash

* Fix `dhall` hash

* Pin to earlier `dhall` for weeder dependency

* Exclude `dhall` to resolve linter failure

The lint job flags `dhall` as needing to be updated from `v1.40.2` to `v1.41.1` but the test job fails with this higher version. Excluding it from the linter since it appears still necessary to support `weeder-2.2.0`

* Use `yesod-routes-flow` in Stackage

* `yesod-routes-flow` package

* Group packages

Co-authored-by: pbrisbin <pbrisbin@users.noreply.github.com>
Co-authored-by: ✿ corey (he/they) <cjohns@freckle.com>
Co-authored-by: ✿ corey (he/they) <corey@x64.co>