Replace example project with real test project
This removes the test project, which aimed to list dependencies typical
of our projects to ensure we were making a coherent snapshot and avoid
problems on use that require fixes.

But it's not exactly right. We almost always go to our most complicated
Haskell project (megarepo) and test the new snapshot with it before
merging anyway. So why not just let CI do that instead?